feature: secured acces to all fields assigned to a single ApplicationUser to that User

See merge request 2023ws-se-pr-group/23ws-se-pr-qse-09!58

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/CalendarReferenceServiceImpl.java
@@ -37,19 +37,33 @@ public class CalendarReferenceServiceImpl implements CalendarReferenceService {
     }
 
     @Override
-    public CalendarReference getFromId(long id) {
+    public CalendarReference getFromId(long id, String username) {
         LOGGER.debug("Get CalendarReference from id {}", id);
-        return calendarReferenceRepository.findById(id).orElseThrow(NotFoundException::new);
+        CalendarReference calendarReference = calendarReferenceRepository
+            .findById(id)
+            .orElseThrow(NotFoundException::new);
+        if (!calendarReference.getUser().getEmail().equals(username)) {
+            throw new AccessDeniedException("Can't fetch Calendars you don't own");
+        }
+        return calendarReference;
     }
 
     @Override
     public CalendarReference add(CalendarReference calendarReference, String username) {
+        if (calendarReference.getId() != null
+            && !calendarReferenceRepository.findById(calendarReference.getId())
+                                           .orElseThrow(NotFoundException::new)
+                                           .getUser()
+                                           .getEmail()
+                                           .equals(username)) {
+            throw new AccessDeniedException("Can't modify Calendars you don't own");
+        }
+
         if (calendarReference.getEnabledDefaultConfigurations() == null) {
             calendarReference.setEnabledDefaultConfigurations(0L);
         }
         LOGGER.debug("Adding CalendarReference {}", calendarReference);
         calendarReference.setToken(generateToken());
-        calendarReference.setEnabledDefaultConfigurations(0L);
         var user = applicationUserRepository.getApplicationUserByEmail(username);
         calendarReference.setUser(user);
         return calendarReferenceRepository.save(calendarReference);
@@ -119,7 +133,14 @@ public class CalendarReferenceServiceImpl implements CalendarReferenceService {
     }
 
     @Override
-    public CalendarReference clonePublicConfig(Long configId, Long calendarId) {
+    public CalendarReference clonePublicConfig(Long configId, Long calendarId, String username) {
+        if (!calendarReferenceRepository.findById(calendarId)
+                                        .orElseThrow(NotFoundException::new)
+                                        .getUser()
+                                        .getEmail()
+                                        .equals(username)) {
+            throw new AccessDeniedException("Can Not assign public Configurations to Calendars you don't own.");
+        }
         Configuration configToAdd = configurationRepository.getReferenceById(configId);
         CalendarReference calendarReference = calendarReferenceRepository.getReferenceById(calendarId);
         if (configId < 0) { // negatives are default configs
@@ -170,12 +191,19 @@ public class CalendarReferenceServiceImpl implements CalendarReferenceService {
 
             return calendarReference;
         } else {
-            throw new AccessDeniedException("Can't add nto public Config");
+            throw new AccessDeniedException("Failed to add public Config.");
         }
     }
 
     @Override
-    public CalendarReference removeConfig(Long configId, Long calendarId) {
+    public CalendarReference removeConfig(Long configId, Long calendarId, String username) {
+        if (!calendarReferenceRepository.findById(calendarId)
+                                        .orElseThrow(NotFoundException::new)
+                                        .getUser()
+                                        .getEmail()
+                                        .equals(username)) {
+            throw new AccessDeniedException("Can Not Remove Configurations from Calendars you don't own.");
+        }
         CalendarReference calendarReference = calendarReferenceRepository.getReferenceById(calendarId);
         if (configId < 0) { // negatives are default configs
             calendarReference.setEnabledDefaultConfigurations(calendarReference.getEnabledDefaultConfigurations() & ~-configId);
@@ -190,8 +218,15 @@ public class CalendarReferenceServiceImpl implements CalendarReferenceService {
     }
 
     @Override
-    public void deleteCalendar(Long id) {
-        LOGGER.debug("Deleting CalendarReference {}", id);
+    public void deleteCalendar(Long id, String username) {
+        LOGGER.debug("Deleting CalendarReference with id: {}, as user: {}", id, username);
+        if (!calendarReferenceRepository.findById(id)
+                                        .orElseThrow(NotFoundException::new)
+                                        .getUser()
+                                        .getEmail()
+                                        .equals(username)) {
+            throw new AccessDeniedException("Can Not Delete Calendars you don't own.");
+        }
         calendarReferenceRepository.deleteById(id);
     }
 }
